Unified CJK characters extraction.
